+++ src/main/java/goorm/kgu/familynote/domain/family/familyQuestion/domain/FamilyQuestion.java
@@ -1,16 +1,14 @@
-package goorm.kgu.familynote.domain.question.familyQuestion.domain;
+package goorm.kgu.familynote.domain.family.familyQuestion.domain;
 
 import goorm.kgu.familynote.common.domain.BaseTimeEntity;
 import goorm.kgu.familynote.domain.family.family.domain.Family;
 import goorm.kgu.familynote.domain.question.baseQuestion.domain.BaseQuestion;
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
