+++ src/main/java/goorm/kgu/familynote/domain/family/familyQuestion/infrastructure/JpaFamilyQuestionRepository.java
@@ -1,12 +1,10 @@
-package goorm.kgu.familynote.domain.question.familyQuestion.infrastructure;
+package goorm.kgu.familynote.domain.family.familyQuestion.infrastructure;
 
-import goorm.kgu.familynote.domain.question.familyQuestion.domain.FamilyQuestion;
+import goorm.kgu.familynote.domain.family.familyQuestion.domain.FamilyQuestion;
 import java.util.List;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
 public interface JpaFamilyQuestionRepository extends JpaRepository<FamilyQuestion, Long> {
